+++ app/src/main/java/com/example/firstapp/Helper/VolleyHelper.kt
@@ -51,7 +51,7 @@ class VolleyHelper (val context : Context){
                 it?.let { jsonArr ->
                     for (i in 0 until jsonArr.length()) {
                         val obj = jsonArr.getJSONObject(i);
-                        val postId = obj.getInt("postId")
+                        val postId = obj.getLong("postId")
                         val title = obj.getString("title")
                         val fileName = obj.getString("file_name")
                         val filePath = obj.getString("path")
